% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papke2022-package.R
\docType{package}
\name{papke2022-package}
\alias{papke2022}
\alias{papke2022-package}
\title{papke2022: Figure Data from Papke et al. (2022)}
\description{
Figure data from Papke et al. (2022) \doi{10.1016/j.neuropharm.2022.109173} obtained by digitization of points from figure images.
}
\author{
\strong{Maintainer}: Ramiro Magno \email{rmagno@pattern.institute} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Authors:
\itemize{
  \item Isabel Duarte \email{iduarte@pattern.institute} (\href{https://orcid.org/0000-0003-0060-2936}{ORCID})
}

Other contributors:
\itemize{
  \item Pattern Institute [copyright holder, funder]
}

}
\keyword{internal}
